# api-docs.yaml
openapi: 3.0.0
paths:
...
  /api/v1/auth/register:
    post:
      operationId: AuthController_register
      summary: Register a new user
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterDto'
      responses:
        '201':
          description: User successfully registered
        '409':
          description: User already exists
      tags: &ref_0
        - auth
  /api/v1/auth/login:
    post:
      operationId: AuthController_login
      summary: Login user
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginDto'
      responses:
        '200':
          description: User successfully logged in
        '401':
          description: Invalid credentials
      tags: *ref_0
  /api/v1/auth/refresh:
    post:
      operationId: AuthController_refresh
      summary: Refresh access token
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenDto'
      responses:
        '200':
          description: Token successfully refreshed
        '401':
          description: Invalid refresh token
      tags: *ref_0
...
  /api/v1/admin/analytics/overview:
    get:
      operationId: AdminAnalyticsController_getOverview
      summary: Get platform overview analytics (Admin only)
      parameters: []
      responses:
        '200':
          description: Overview analytics retrieved successfully
      tags: &ref_1
        - admin-analytics
      security: &ref_2
        - bearer: []
  /api/v1/admin/analytics/users:
    get:
      operationId: AdminAnalyticsController_getUserAnalytics
      summary: Get user analytics (Admin only)
      parameters: []
      responses:
        '200':
          description: User analytics retrieved successfully
      tags: *ref_1
      security: *ref_2
  /api/v1/admin/analytics/products:
    get:
      operationId: AdminAnalyticsController_getProductAnalytics
      summary: Get product analytics (Admin only)
      parameters: []
      responses:
        '200':
          description: Product analytics retrieved successfully
      tags: *ref_1
      security: *ref_2
  /api/v1/admin/analytics/chats:
    get:
      operationId: AdminAnalyticsController_getChatAnalytics
      summary: Get chat analytics (Admin only)
      parameters: []
      responses:
        '200':
          description: Chat analytics retrieved successfully
      tags: *ref_1
      security: *ref_2
  /api/v1/admin/analytics/engagement:
    get:
      operationId: AdminAnalyticsController_getEngagementAnalytics
      summary: Get engagement analytics (Admin only)
      parameters: []
      responses:
        '200':
          description: Engagement analytics retrieved successfully
      tags: *ref_1
      security: *ref_2
  /api/v1/admin/analytics/reports:
    get:
      operationId: AdminAnalyticsController_getTrustSafetyAnalytics
      summary: Get trust & safety analytics (Admin only)
      parameters: []
      responses:
        '200':
          description: Trust & safety analytics retrieved successfully
      tags: *ref_1
      security: *ref_2
  /api/v1/admin/analytics/system:
    get:
      operationId: AdminAnalyticsController_getSystemAnalytics
      summary: Get system & security analytics (Admin only)
      parameters: []
      responses:
        '200':
          description: System analytics retrieved successfully
      tags: *ref_1
      security: *ref_2
  /api/v1/admin/analytics/test:
    get:
      operationId: AdminAnalyticsController_test
      summary: Test endpoint
      parameters: []
      responses:
        '200':
          description: Test successful
      tags: *ref_1
      security: *ref_2
...
  /api/v1/users/test:
    get:
      operationId: UsersController_test
      summary: Test endpoint
      parameters: []
      responses:
        '200':
          description: Test successful
      tags: *ref_3
      security: *ref_4
...
  /api/v1/categories:
    get:
      operationId: CategoriesController_findAll
      summary: Get all categories
      parameters: []
      responses:
        '200':
          description: Categories retrieved successfully
      tags: &ref_5
        - categories
    post:
      operationId: CategoriesController_create
      summary: Create a new category (Admin only)
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCategoryDto'
      responses:
        '201':
          description: Category created successfully
        '409':
          description: Category already exists
      tags: *ref_5
      security:
        - bearer: []
...
  /api/v1/categories/{id}:
    get:
      operationId: CategoriesController_findById
      summary: Get category by ID
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Category retrieved successfully
        '404':
          description: Category not found
      tags: *ref_5
    put:
      operationId: CategoriesController_update
      summary: Update a category (Admin only)
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCategoryDto'
      responses:
        '200':
          description: Category updated successfully
        '404':
          description: Category not found
      tags: *ref_5
      security:
        - bearer: []
    delete:
      operationId: CategoriesController_delete
      summary: Delete a category (Admin only)
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Category deleted successfully
        '404':
          description: Category not found
      tags: *ref_5
      security:
        - bearer: []
  /api/v1/categories/slug/{slug}:
    get:
      operationId: CategoriesController_findBySlug
      summary: Get category by slug
      parameters:
        - name: slug
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Category retrieved successfully
        '404':
          description: Category not found
      tags: *ref_5
  /api/v1/categories/subcategories:
    post:
      operationId: CategoriesController_createSubcategory
      summary: Create a new subcategory (Admin only)
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSubcategoryDto'
      responses:
        '201':
          description: Subcategory created successfully
        '409':
          description: Subcategory already exists
      tags: *ref_5
      security:
        - bearer: []
  /api/v1/categories/subcategories/{id}:
    delete:
      operationId: CategoriesController_deleteSubcategory
      summary: Delete a subcategory (Admin only)
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Subcategory deleted successfully
        '404':
          description: Subcategory not found
      tags: *ref_5
      security:
        - bearer: []
...
  /api/v1/reports:
    post:
      operationId: ReportsController_create
      summary: Create a new report
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReportDto'
      responses:
        '201':
          description: Report created successfully
        '400':
          description: Invalid request
      tags: &ref_13
        - reports
      security: &ref_14
        - bearer: []
    get:
      operationId: ReportsController_findAll
      summary: Get all reports (Admin only)
      parameters:
        - name: status
          required: true
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Reports retrieved successfully
      tags: *ref_13
      security: *ref_14
...
  /api/v1/reports/test:
    get:
      operationId: ReportsController_test
      summary: Test endpoint
      parameters: []
      responses:
        '200':
          description: Test successful
      tags: *ref_13
      security: *ref_14
...
info:
  title: memi.az API
  description: Recommerce Platform Backend API
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    RegisterDto:
      type: object
      properties: {}
    LoginDto:
      type: object
      properties: {}
    RefreshTokenDto:
      type: object
      properties: {}
...
    CreateCategoryDto:
      type: object
      properties: {}
    UpdateCategoryDto:
      type: object
      properties: {}
    CreateSubcategoryDto:
      type: object
      properties: {}
...
    CreateReportDto:
      type: object
      properties: {}
...